Split vnf.c into common, LTE and NR implementations
